ensure bools are converted to int

+++ contrib/reinvent_plugins/components/NIBRSubstructureFilters/comp_nibr_substructure_filter.py
@@ -53,6 +53,8 @@ class NIBRSubstructureFilters:
 
             # SubstructureMatches, Min_N_O_filter, Frac_N_O, Covalent,
             # SpecialMol, SeverityScore
-            scores.append(np.array([entry.SeverityScore < cutoff for entry in nibr_scores]))
+            scores.append(
+                np.array([entry.SeverityScore < cutoff for entry in nibr_scores], dtype=int)
+            )
 
         return ComponentResults(scores)
